requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# GitHub Integration
PyGithub>=2.1.0

# Code Analysis
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0

# Configuration & Data
pyyaml>=6.0
pydantic>=2.5.0

# HTTP Client
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
